CARDS-1132: New charting component used by every chart widget Add the line graph, and break out the statistics into its own component
